move i18n tests to separate file
