+++ tests/py/test_utils.py
@@ -5,13 +5,12 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 from datetime import datetime
 from datetime import timedelta
 
-from mangopay.utils import Money
 from markupsafe import escape
 from pando.http.response import Response
 
 from liberapay import utils
 from liberapay.testing import Harness
-from liberapay.utils import i18n, markdown, b64encode_s, b64decode_s
+from liberapay.utils import markdown, b64encode_s, b64decode_s
 from liberapay.wireup import CSP
 
 
@@ -97,24 +96,6 @@ class Tests(Harness):
     def test_not_expired(self):
         expiration = datetime.utcnow() + timedelta(days=100)
         assert not utils.is_card_expired(expiration.year, expiration.month)
-
-    # i18n
-    # ====
-
-    def test_format_currency_without_trailing_zeroes(self):
-        expected = '$16'
-        actual = i18n.LOCALE_EN.format_money(Money(16, 'USD'), trailing_zeroes=False)
-        assert actual == expected
-
-    def test_format_currency_defaults_to_trailing_zeroes(self):
-        expected = '$16.00'
-        actual = i18n.LOCALE_EN.format_money(Money(16, 'USD'))
-        assert actual == expected
-
-    def test_locales_share_message_keys(self):
-        msgkey1 = self.website.locales['de'].catalog['Save'].id
-        msgkey2 = self.website.locales['fr'].catalog['Save'].id
-        assert id(msgkey1) == id(msgkey2)
 
     # Markdown
     # ========
